First steps into better field generation
Implements a look-around for each opportunity to place a mine to avoid enclosing any adjacent fieldElem with mines by placing said mine in the currently opportune spot.
Needs testing regarding if this might have broken that all mines kann be placed on the given field and/or a way to determine this before-hand.
The Bunch class got some changes as well to try to allow  "bunchObj"."someItem" to work just as "bunchObj"["someItem"] does.