Revisions to options-file and object-structure
Improved reading from fieldOptions.txt and improved documentation in it as well.
Moved the FIeld-object to be part of the window-object (as is the plan for any- and everything else) too allow the internal workings of the application to reach the user (for example through messageBoxes as in the Field-class).
Need to work on the field generation.
-> Fields where area == mines seem bugged